+++ src/components/imageOrGalleryRoute.jsx
